moved category and workflowstage queries to their own modules
