+++ src/seis_lab_data/db/queries/surveyrelatedrecords.py
@@ -1,5 +1,4 @@
 import logging
-import uuid
 
 import shapely
 from sqlalchemy.orm import selectinload
@@ -319,101 +318,3 @@ async def list_survey_related_record_subject_records(
         .options(selectinload(models.SurveyRelatedRecord.workflow_stage))
     )
     return (await session.exec(statement.offset(offset).limit(limit))).all()
-
-
-async def list_dataset_categories(
-    session: AsyncSession,
-    limit: int = 20,
-    offset: int = 0,
-    include_total: bool = False,
-    order_by_clause=models.DatasetCategory.name["en"].astext,
-) -> tuple[list[models.DatasetCategory], int | None]:
-    # NOTE: limit, offset and order_by are applied only when asking the
-    # session to exec because we want to reuse the statement later, to count
-    # total number of records
-    statement = select(models.DatasetCategory)
-    items = (
-        await session.exec(
-            statement.limit(limit).offset(offset).order_by(order_by_clause)
-        )
-    ).all()
-
-    num_total = (
-        await _get_total_num_records(session, statement) if include_total else None
-    )
-    return items, num_total
-
-
-async def collect_all_dataset_categories(
-    session: AsyncSession, order_by_clause=models.DatasetCategory.name["en"].astext
-) -> list[models.DatasetCategory]:
-    _, num_total = await list_dataset_categories(session, limit=1, include_total=True)
-    items, _ = await list_dataset_categories(
-        session, limit=num_total, include_total=False, order_by_clause=order_by_clause
-    )
-    return items
-
-
-async def get_dataset_category(
-    session: AsyncSession,
-    dataset_category_id: uuid.UUID,
-) -> models.DatasetCategory | None:
-    return await session.get(models.DatasetCategory, dataset_category_id)
-
-
-async def get_dataset_category_by_english_name(
-    session: AsyncSession, name: str
-) -> models.DatasetCategory | None:
-    statement = select(models.DatasetCategory).where(
-        models.DatasetCategory.name["en"].astext == name
-    )
-    return (await session.exec(statement)).first()
-
-
-async def list_workflow_stages(
-    session: AsyncSession,
-    limit: int = 20,
-    offset: int = 0,
-    include_total: bool = False,
-    order_by_clause=models.WorkflowStage.name["en"].astext,
-) -> tuple[list[models.WorkflowStage], int | None]:
-    statement = select(models.WorkflowStage)
-
-    # NOTE: limit, offset and order_by are applied only when asking the
-    # session to exec because we want to reuse the statement later, to count
-    # total number of records
-    items = (
-        await session.exec(
-            statement.offset(offset).limit(limit).order_by(order_by_clause)
-        )
-    ).all()
-    num_total = (
-        await _get_total_num_records(session, statement) if include_total else None
-    )
-    return items, num_total
-
-
-async def collect_all_workflow_stages(
-    session: AsyncSession, order_by_clause=models.WorkflowStage.name["en"].astext
-) -> list[models.WorkflowStage]:
-    _, num_total = await list_workflow_stages(session, limit=1, include_total=True)
-    items, _ = await list_workflow_stages(
-        session, limit=num_total, include_total=False, order_by_clause=order_by_clause
-    )
-    return items
-
-
-async def get_workflow_stage(
-    session: AsyncSession,
-    workflow_stage_id: uuid.UUID,
-) -> models.WorkflowStage | None:
-    return await session.get(models.WorkflowStage, workflow_stage_id)
-
-
-async def get_workflow_stage_by_english_name(
-    session: AsyncSession, name: str
-) -> models.WorkflowStage | None:
-    statement = select(models.WorkflowStage).where(
-        models.WorkflowStage.name["en"].astext == name
-    )
-    return (await session.exec(statement)).first()
