+++ src/seis_lab_data/db/queries/workflowstages.py
@@ -1,0 +1,59 @@
+import logging
+import uuid
+
+from sqlmodel.ext.asyncio.session import AsyncSession
+from sqlmodel import select
+
+from ...db import models
+from .common import _get_total_num_records
+
+logger = logging.getLogger(__name__)
+
+
+async def list_workflow_stages(
+    session: AsyncSession,
+    limit: int = 20,
+    offset: int = 0,
+    include_total: bool = False,
+    order_by_clause=models.WorkflowStage.name["en"].astext,
+) -> tuple[list[models.WorkflowStage], int | None]:
+    statement = select(models.WorkflowStage)
+
+    # NOTE: limit, offset and order_by are applied only when asking the
+    # session to exec because we want to reuse the statement later, to count
+    # total number of records
+    items = (
+        await session.exec(
+            statement.offset(offset).limit(limit).order_by(order_by_clause)
+        )
+    ).all()
+    num_total = (
+        await _get_total_num_records(session, statement) if include_total else None
+    )
+    return items, num_total
+
+
+async def collect_all_workflow_stages(
+    session: AsyncSession, order_by_clause=models.WorkflowStage.name["en"].astext
+) -> list[models.WorkflowStage]:
+    _, num_total = await list_workflow_stages(session, limit=1, include_total=True)
+    items, _ = await list_workflow_stages(
+        session, limit=num_total, include_total=False, order_by_clause=order_by_clause
+    )
+    return items
+
+
+async def get_workflow_stage(
+    session: AsyncSession,
+    workflow_stage_id: uuid.UUID,
+) -> models.WorkflowStage | None:
+    return await session.get(models.WorkflowStage, workflow_stage_id)
+
+
+async def get_workflow_stage_by_english_name(
+    session: AsyncSession, name: str
+) -> models.WorkflowStage | None:
+    statement = select(models.WorkflowStage).where(
+        models.WorkflowStage.name["en"].astext == name
+    )
+    return (await session.exec(statement)).first()
